miner_restarter/app/requirements.txt
fastapi[standard]
uvicorn>=0.24.0
pydantic>=2.4.2
aiohttp
asyncssh
python-dotenv
pydantic-settings
requests


